services/server/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
flask-migrate==2.7.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
